+++ src/main/java/interface_adapter/office/SimulateController.java
@@ -19,7 +19,7 @@ public class SimulateController {
     /**
      * Executes the Simulate Use Case.
      */
-    public void execute(int currentDay, int currentBalance, int pastCustomerCount) {
+    public void execute(int currentDay, double currentBalance, int pastCustomerCount) {
         final SimulateInputData simulateInputData = new SimulateInputData(currentDay, currentBalance, pastCustomerCount);
 
         simulateUseCaseInteractor.execute(simulateInputData);
