+++ src/main/java/use_case/simulate/SimulateOutputData.java
@@ -2,10 +2,10 @@ package use_case.simulate;
 
 public class SimulateOutputData {
     private final int currentDay;
-    private final int currentBalance;
+    private final double currentBalance;
     private final int pastCustomerCount;
 
-    public SimulateOutputData(int currentDay, int currentBalance, int pastCustomerCount) {
+    public SimulateOutputData(int currentDay, double currentBalance, int pastCustomerCount) {
         this.currentDay = currentDay;
         this.currentBalance = currentBalance;
         this.pastCustomerCount = pastCustomerCount;
@@ -14,7 +14,7 @@ public class SimulateOutputData {
     public int getCurrentDay() {
         return currentDay;
     }
-    public int getCurrentBalance() {
+    public double getCurrentBalance() {
         return currentBalance;
     }
     public int getPastCustomerCount() {return pastCustomerCount;}
